% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{add_rcrd_log}
\alias{add_rcrd_log}
\title{Log sequences added to the SQL database in the restez path}
\usage{
add_rcrd_log(df)
}
\arguments{
\item{df}{sequence records, data.frame}
}
\description{
This function is called whenever sequences have been
successfully added to the nucleotide SQL database. Row entries are added to
'add_lot.tsv' in the user's restez path containing the sequence accession,
version and GB release numbers as well as the time of successfuly adding.
The log is to help users keep track of when sequences have been added.
}
\seealso{
Other private: \code{\link{check_connection}},
  \code{\link{cleanup}}, \code{\link{connection_get}},
  \code{\link{custom_download}},
  \code{\link{db_nrows_get}}, \code{\link{dir_size}},
  \code{\link{dwnld_path_get}},
  \code{\link{dwnld_rcrd_log}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_keyword}},
  \code{\link{extract_clean_sequence}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_inforecpart}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_seqrecpart}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{flatfile_read}},
  \code{\link{gb_df_create}}, \code{\link{gb_df_generate}},
  \code{\link{gb_sql_add}}, \code{\link{gb_sql_query}},
  \code{\link{gbrelease_check}},
  \code{\link{gbrelease_log}},
  \code{\link{identify_downloadable_files}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
  \code{\link{last_entry_get}},
  \code{\link{message_missing}}, \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_path_check}},
  \code{\link{restez_rl}}, \code{\link{setup}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
