% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{check_status}
\alias{check_status}
\title{Check status}
\usage{
check_status()
}
\description{
Report to console current setup status of restez.
Determines if the restez path is set, how many downloaded files
there are, if there is an SQL database.
}
\examples{
library(restez)
fp <- tempdir()
set_restez_path(filepath = fp)
create_demo_database()
check_status()
delete_database()
check_status()
}
\seealso{
Other setup: \code{\link{create_database}},
  \code{\link{create_demo_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}},
  \code{\link{unset_restez_path}}
}
