% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{connection_get}
\alias{connection_get}
\title{Retrieve restez connection}
\usage{
connection_get()
}
\value{
connection
}
\description{
Safely acquire the restez connection. Raises error if no
connection set.
}
\seealso{
Other setup: \code{\link{db_create}},
  \code{\link{db_delete}}, \code{\link{db_download}},
  \code{\link{demo_db_create}},
  \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
