% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-get-tools.R
\name{count_db_ids}
\alias{count_db_ids}
\title{Return the number of ids}
\usage{
count_db_ids(db = "nucleotide")
}
\arguments{
\item{db}{character, database name}
}
\value{
integer
}
\description{
Return the number of ids in a user's restez database.
}
\details{
Requires an open connection. If no connection or db 0 is returned.
}
\examples{
library(restez)
restez_path_set(filepath = tempdir())
restez_connect()
demo_db_create()
(count_db_ids())

# delete demo after example
db_delete(everything = TRUE)
}
\seealso{
Other get: \code{\link{gb_definition_get}},
  \code{\link{gb_fasta_get}},
  \code{\link{gb_organism_get}},
  \code{\link{gb_record_get}},
  \code{\link{gb_sequence_get}},
  \code{\link{gb_version_get}}, \code{\link{is_in_db}},
  \code{\link{list_db_ids}}
}
