% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_database}
\alias{create_database}
\title{Create database}
\usage{
create_database(db_type = "nucleotide", overwrite = FALSE)
}
\arguments{
\item{db_type}{character, database type}

\item{overwrite}{T/F, overwrite files already in database?}
}
\description{
Checks for downloaded .seq.tar files,
decompresses and then adds the files to a local SQL
database.
}
