% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_demo_database}
\alias{create_demo_database}
\title{Create demo database}
\usage{
create_demo_database(db_type = "nucleotide", n = 100)
}
\arguments{
\item{db_type}{character, database type}

\item{n}{integer, number of mock sequences}
}
\description{
Creates a local mock SQL database
from package test data for demonstration purposes.
No internet connection required.
}
\examples{
library(restez)
# set the restez path to a temporary dir
set_restez_path(filepath = tempdir())
# create demo database
create_demo_database(n = 100)
# in the demo, IDs are 'demo_1', 'demo_2' ...
(get_sequence(id = 'demo_1'))

# Be careful with the database as it build cummulatively.
#  if we running create_demo_database/create_database
#  will append new entries to the preexisitng database
create_demo_database(n = 100)
# Leading to multiple entries if IDs are the same
(get_sequence(id = 'demo_1'))

# To resovle this issue, always delete a demo database
#  after an example
delete_database()
}
\seealso{
Other setup: \code{\link{check_status}},
  \code{\link{create_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}},
  \code{\link{unset_restez_path}}
}
