% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_download}
\alias{db_download}
\title{Download database}
\usage{
db_download(db = "nucleotide", overwrite = FALSE, preselection = NULL)
}
\arguments{
\item{db}{Database type, only 'nucleotide' currently available.}

\item{overwrite}{T/F, overwrite pre-existing downloaded files?}

\item{preselection}{character of user input}
}
\value{
T/F, if all files download correctly, TRUE else FALSE.
}
\description{
Download .seq.tar files from the latest GenBank release. The
user interactively selects the parts of GenBank to download (e.g. primates,
plants, bacteria ...)
}
\details{
The downloaded files will appear in the restez filepath under downloads.
}
\examples{
\dontrun{
library(restez)
restez_path_set(filepath = 'path/for/downloads')
db_download()
}
}
\seealso{
Other setup: \code{\link{connection_get}},
  \code{\link{db_create}}, \code{\link{db_delete}},
  \code{\link{demo_db_create}},
  \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
