% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{demo_db_create}
\alias{demo_db_create}
\title{Create demo database}
\usage{
demo_db_create(db_type = "nucleotide", n = 100)
}
\arguments{
\item{db_type}{character, database type}

\item{n}{integer, number of mock sequences}
}
\description{
Creates a local mock SQL database
from package test data for demonstration purposes.
No internet connection required.
}
\examples{
library(restez)
# set the restez path to a temporary dir
restez_path_set(filepath = tempdir())
# create demo database
demo_db_create(n = 100)
# in the demo, IDs are 'demo_1', 'demo_2' ...
(gb_sequence_get(id = 'demo_1'))

# Be careful with the database as it build cummulatively.
#  if we running demo_db_create/db_create
#  will append new entries to the preexisitng database
demo_db_create(n = 100)
# Leading to multiple entries if IDs are the same
(gb_sequence_get(id = 'demo_1'))

# To resovle this issue, always delete a demo database
#  after an example
db_delete()
}
\seealso{
Other setup: \code{\link{db_create}},
  \code{\link{db_delete}}, \code{\link{gb_download}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{status_check}}
}
