% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{disconnect}
\alias{disconnect}
\title{Disconnect from a database}
\usage{
disconnect(connection)
}
\description{
Safely disconnect form a connection
}
\seealso{
Other setup: \code{\link{connect}},
  \code{\link{db_create}}, \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
