% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{download_genbank}
\alias{download_genbank}
\title{Download GenBank}
\usage{
download_genbank()
}
\description{
Download .seq files from the latest GenBank
release. The user interacitvely selects the parts of
GenBank to download (e.g. primates, plants, bacteria ...)
}
\details{
The downloaded files will appear in the restez filepath under
downloads.
The downloaded files will be .tar compressed and will need
extraction for inspection. (setup() does this itself.)
Warning: this function will overwrite pre-exisitng files.
}
