% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{dwnld_rcrd_log}
\alias{dwnld_rcrd_log}
\title{Log a downloaded file in the restez path}
\usage{
dwnld_rcrd_log(fl)
}
\arguments{
\item{fl}{file name, character}
}
\description{
This function is called whenever a file is successfully
downloaded. A row entry is added to the 'download_log.tsv' in the user's
restez path containing the file name, the GB release number and the time of
successfuly download. The log is to help users keep track of when they
downloaded files and to determine if the downloaded files are out of date.
}
\seealso{
Other private: \code{\link{add_rcrd_log}},
  \code{\link{check_connection}}, \code{\link{cleanup}},
  \code{\link{connection_get}},
  \code{\link{custom_download}},
  \code{\link{db_nrows_get}}, \code{\link{dir_size}},
  \code{\link{dwnld_path_get}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_keyword}},
  \code{\link{extract_clean_sequence}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_inforecpart}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_seqrecpart}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{flatfile_read}},
  \code{\link{gb_df_create}}, \code{\link{gb_df_generate}},
  \code{\link{gb_sql_add}}, \code{\link{gb_sql_query}},
  \code{\link{gbrelease_check}},
  \code{\link{gbrelease_log}},
  \code{\link{identify_downloadable_files}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
  \code{\link{last_entry_get}},
  \code{\link{message_missing}}, \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_path_check}},
  \code{\link{restez_rl}}, \code{\link{setup}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
