% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentrez-wrappers.R
\name{entrez_fetch}
\alias{entrez_fetch}
\title{Entrez fetch}
\usage{
entrez_fetch(db, id = NULL, rettype, retmode = "", ...)
}
\arguments{
\item{db}{character, name of the database}

\item{id}{vector, unique ID(s) for record(s)}

\item{rettype}{character, data format}

\item{retmode}{character, data mode}

\item{...}{Arguments to be passed on to rentrez}
}
\value{
character string containing the file created
}
\description{
Wrapper for rentrez::entrez_fetch.
}
\details{
Attempts to first search local database with user-sepcified
parameters, if the record is missing in the database, the function then
calls rentrez::entrez_fetch to search GenBank remotely.
}
\seealso{
\code{\link[rentrez]{entrez_fetch}}
}
