% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tools.R
\name{extract_by_keyword}
\alias{extract_by_keyword}
\title{Extract by keyword}
\usage{
extract_by_keyword(record, keyword, end_pattern = "\\n")
}
\arguments{
\item{record}{GenBank record in text format, character}

\item{keyword}{Keyword, character}

\item{end_pattern}{REGEX pattern indicating the point to
stop extraction, character}
}
\value{
character or NULL
}
\description{
Search through GenBank record for a keyword and
return text up to the end_pattern.
}
\details{
The keyword should be any of the capitalized elements
in a GenBank record (e.g. LOCUS, DESCRIPTION, ACCESSION).
The end_pattern depends on how much of the selected element
a user wants returned. By default, the extraction will stop
at the next newline.
If keyword or end pattern not found, returns NULL.
}
\seealso{
Other private: \code{\link{check_connection}},
  \code{\link{cleanup}}, \code{\link{connection_get}},
  \code{\link{custom_download}},
  \code{\link{dwnld_path_get}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{flatfile_read}},
  \code{\link{gb_df_create}}, \code{\link{gb_df_generate}},
  \code{\link{gb_sql_add}}, \code{\link{gb_sql_query}},
  \code{\link{identify_downloadable_files}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_path_check}},
  \code{\link{restez_rl}},
  \code{\link{save_download_details}}, \code{\link{setup}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
