% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tools.R
\name{extract_definition}
\alias{extract_definition}
\title{Extract definition}
\usage{
extract_definition(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return definition from GenBank record
}
