% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tools.R
\name{extract_organism}
\alias{extract_organism}
\title{Extract organism}
\usage{
extract_organism(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return organism name from GenBank record
}
