% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tools.R
\name{extract_sequence}
\alias{extract_sequence}
\title{Extract sequence}
\usage{
extract_sequence(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return sequecne from GenBank record
}
