% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tools.R
\name{extract_version}
\alias{extract_version}
\title{Extract version}
\usage{
extract_version(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return version ID from GenBank record
}
