% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{file_download}
\alias{file_download}
\title{Download a file}
\usage{
file_download(fl, overwrite = FALSE)
}
\arguments{
\item{fl}{character, base filename (e.g. gbpri9.seq) to be
downloaded}

\item{overwrite}{T/F}
}
\value{
T/F
}
\description{
Download a GenBank .seq.tar file. Check
the file has downloaded properly. If not, returns FALSE.
If overwrite is true, any previous file will be overwritten.
}
\seealso{
Other private: \code{\link{check_connection}},
  \code{\link{cleanup}}, \code{\link{custom_download}},
  \code{\link{dwnld_path_get}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_keyword}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{flatfile_read}}, \code{\link{gb_df_create}},
  \code{\link{gb_df_generate}}, \code{\link{gb_sql_add}},
  \code{\link{gb_sql_query}},
  \code{\link{identify_downloadable_files}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_path_check}},
  \code{\link{restez_rl}},
  \code{\link{save_download_details}}, \code{\link{setup}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
