% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gb-setup-tools.R
\name{gb_df_create}
\alias{gb_df_create}
\title{Create GenBank data.frame}
\usage{
gb_df_create(accessions, versions, organisms, definitions, sequences, records)
}
\arguments{
\item{accessions}{character, vector of accessions}

\item{versions}{character, vector of accessions + versions}

\item{organisms}{character, vector of organism names}

\item{definitions}{character, vector of sequence definitions}

\item{sequences}{character, vector of sequences}

\item{records}{character, vector of GenBank records in text format}
}
\value{
data.frame
}
\description{
Make data.frame from columns vectors for
nucleotide entries. As part of gb_df_generate().
}
\seealso{
Other private: \code{\link{check_connection}},
  \code{\link{cleanup}}, \code{\link{custom_download}},
  \code{\link{dwnld_path_get}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_keyword}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{flatfile_read}},
  \code{\link{gb_df_generate}}, \code{\link{gb_sql_add}},
  \code{\link{gb_sql_query}},
  \code{\link{identify_downloadable_files}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_path_check}},
  \code{\link{restez_rl}},
  \code{\link{save_download_details}}, \code{\link{setup}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
