% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{gb_download}
\alias{gb_download}
\title{Download GenBank}
\usage{
gb_download(overwrite = FALSE)
}
\arguments{
\item{overwrite}{T/F, overwrite pre-existing downloaded files?}
}
\description{
Download .seq.tar files from the latest GenBank
release. The user interacitvely selects the parts of
GenBank to download (e.g. primates, plants, bacteria ...)
}
\details{
The downloaded files will appear in the restez filepath under
downloads.
}
\examples{
\dontrun{
library(restez)
restez_path_set(filepath = 'path/for/downloads')
gb_download()
}
}
\seealso{
Other setup: \code{\link{db_create}},
  \code{\link{db_delete}}, \code{\link{demo_db_create}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{status_check}}
}
