% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-tools.R
\name{generate_dataframe}
\alias{generate_dataframe}
\title{Generate records data.frame}
\usage{
generate_dataframe(records)
}
\arguments{
\item{records}{list of GenBank records in text format}
}
\value{
data.frame
}
\description{
For a list of records, construct a data.frame
for insertion into SQL database.
}
\details{
The resulting data.frame has five columns: accession,
organism, raw_definition, raw_sequence, raw_record.
The prefix 'raw_' indicates the data has been covnerted to the
raw format, see ?charToRaw, in order to save on RAM.
The raw_record contains the entire GenBank record in text format.
}
