% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_definition}
\alias{get_definition}
\title{Get definition}
\usage{
get_definition(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
list of definitions
}
\description{
Return the definition line
for an accession ID.
}
\seealso{
Other get: \code{\link{get_organism}},
  \code{\link{get_record}}, \code{\link{get_sequence}},
  \code{\link{list_db_ids}}
}
