% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_fasta}
\alias{get_fasta}
\title{Get fasta}
\usage{
get_fasta(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
named vector of fasta sequences
}
\description{
Get sequence and definition data
in FASTA format.
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database()
fasta <- get_fasta(id = 'demo_1')
fastas <- get_fasta(id = c('demo_1', 'demo_2'))

# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_organism}}, \code{\link{get_record}},
  \code{\link{get_sequence}}, \code{\link{get_version}},
  \code{\link{is_in_db}}, \code{\link{list_db_ids}}
}
