% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_organism}
\alias{get_organism}
\title{Get organism}
\usage{
get_organism(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
list of definitions
}
\description{
Return the organism name
for an accession ID.
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_record}}, \code{\link{get_sequence}},
  \code{\link{list_db_ids}}
}
