% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_record}
\alias{get_record}
\title{Get record}
\usage{
get_record(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
named vector of records, if no results found NULL
}
\description{
Return the entire GenBank record
for an accession ID.
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database()
(rec <- get_record(id = 'demo_1'))
(recs <- get_record(id = c('demo_1', 'demo_2')))


# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_fasta}}, \code{\link{get_organism}},
  \code{\link{get_sequence}}, \code{\link{get_version}},
  \code{\link{is_in_db}}, \code{\link{list_db_ids}}
}
