% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_sequence}
\alias{get_sequence}
\title{Get sequence}
\usage{
get_sequence(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
list of sequences
}
\description{
Return the sequence(s) for a record(s)
from the accession ID(s).
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database()
seq <- get_sequence(id = 'demo_1')
seqs <- get_sequence(id = c('demo_1', 'demo_2'))

# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_organism}}, \code{\link{get_record}},
  \code{\link{list_db_ids}}
}
