% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-tools.R
\name{get_sequence_from_id}
\alias{get_sequence_from_id}
\title{Get sequence from ID}
\usage{
get_sequence_from_id(id, filepath)
}
\arguments{
\item{id}{sequence accession ID}

\item{filepath}{Database filepath}
}
\value{
character
}
\description{
Return the sequence for a record
from the accession ID.
}
