% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_version}
\alias{get_version}
\title{Get version}
\usage{
get_version(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
named vector of versions, if no results found NULL
}
\description{
Return the accession version
for an accession ID.
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database()
(ver <- get_version(id = 'demo_1'))
(vers <- get_version(id = c('demo_1', 'demo_2')))


# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_fasta}}, \code{\link{get_organism}},
  \code{\link{get_record}}, \code{\link{get_sequence}},
  \code{\link{is_in_db}}, \code{\link{list_db_ids}}
}
