% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{identify_downloadable_files}
\alias{identify_downloadable_files}
\title{Identify downloadable files}
\usage{
identify_downloadable_files(release_notes)
}
\arguments{
\item{release_notes}{character, GenBank release notes}
}
\value{
data.frame
}
\description{
Searches through the release notes
for a GenBank release to find all listed .seq files.
Returns a data.frame for all .seq files and their
description.
}
