% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{identify_latest_genbank_release_notes}
\alias{identify_latest_genbank_release_notes}
\title{Identify the latest GenBank Release Notes}
\usage{
identify_latest_genbank_release_notes()
}
\value{
character
}
\description{
Searches through all release notes
to find the latest. Returns the entire release notes.
}
