% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-get-tools.R
\name{is_in_db}
\alias{is_in_db}
\title{Is in db}
\usage{
is_in_db(id, db = "nucleotide")
}
\arguments{
\item{id}{character, sequence accession ID(s)}

\item{db}{character, database name}
}
\value{
named vector of booleans
}
\description{
Determine whether an id(s)
is/are present in a database.
}
\examples{
library(restez)
# set the restez path to a temporary dir
restez_path_set(filepath = tempdir())
restez_connect()
# create demo database
demo_db_create(n = 100)
# in the demo, IDs are 'demo_1', 'demo_2' ...
ids <- c('thisisnotanid', 'demo_1', 'demo_2')
(is_in_db(id = ids))


# delete demo after example
restez_disconnect()
db_delete()
}
\seealso{
Other get: \code{\link{gb_definition_get}},
  \code{\link{gb_fasta_get}},
  \code{\link{gb_organism_get}},
  \code{\link{gb_record_get}},
  \code{\link{gb_sequence_get}},
  \code{\link{gb_version_get}}, \code{\link{list_db_ids}}
}
