% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{list_db_ids}
\alias{list_db_ids}
\title{List database IDs}
\usage{
list_db_ids(db)
}
\arguments{
\item{db}{character, database name}
}
\value{
vector of characters
}
\description{
Return a vector of all IDs in
a database.
}
\details{
Warning: can return very large vectors
for large databases.
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_organism}}, \code{\link{get_record}},
  \code{\link{get_sequence}}
}
