% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-tools.R
\name{print.status}
\alias{print.status}
\title{Print method for status class}
\usage{
\method{print}{status}(x, ...)
}
\arguments{
\item{x}{Status object}

\item{...}{Other arguments (not used by this function)}
}
\description{
Prints to screen the three sections of the status class. Not
meant to be used interactively.
}
