% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-tools.R
\name{read_records}
\alias{read_records}
\title{Read records}
\usage{
read_records(filepath)
}
\arguments{
\item{filepath}{Path to .seq file}
}
\value{
list of GenBank records in text format
}
\description{
Read records from a .seq file.
}
