\name{record}
\alias{record}
\docType{data}
\title{
Example GenBank record
}
\description{
Example GenBank record in text format for demonstration purposes.
}
\usage{data("record")}
\format{
  A large character object containing record information and DNA sequence.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/AY952423.1}
}
\references{
GenBank
}
\examples{
data(record)
cat(record)
}
\keyword{datasets}
