% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_connect}
\alias{restez_connect}
\title{Connect to the restez database}
\usage{
restez_connect()
}
\description{
Safely creates returns a connection to the local database. If
database connection cannot be made, an error is returned.
}
\seealso{
Other setup: \code{\link{connection_get}},
  \code{\link{db_create}}, \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
