% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_disconnect}
\alias{restez_disconnect}
\title{Disconnect from restez database}
\usage{
restez_disconnect()
}
\description{
Safely disconnect from the restez connection
}
\seealso{
Other setup: \code{\link{connection_get}},
  \code{\link{db_create}}, \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{restez_connect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
