% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_path_check}
\alias{restez_path_check}
\title{Check restez filepath}
\usage{
restez_path_check()
}
\description{
Raises error if restez path does
not exist.
}
\seealso{
Other private: \code{\link{check_connection}},
  \code{\link{cleanup}}, \code{\link{custom_download}},
  \code{\link{dwnld_path_get}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_keyword}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{flatfile_read}},
  \code{\link{gb_df_create}}, \code{\link{gb_df_generate}},
  \code{\link{gb_sql_add}}, \code{\link{gb_sql_query}},
  \code{\link{identify_downloadable_files}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_rl}},
  \code{\link{save_download_details}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
