% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_path_get}
\alias{restez_path_get}
\title{Get restez path}
\usage{
restez_path_get()
}
\value{
character
}
\description{
Return filepath to where the restez
database is stored.
}
\examples{
library(restez)
# set a restez path with a tempdir
restez_path_set(filepath = tempdir())
# check what the set path is
(restez_path_get())
}
\seealso{
Other setup: 
\code{\link{restez_path_set}()},
\code{\link{restez_path_unset}()},
\code{\link{restez_ready}()},
\code{\link{restez_status}()}
}
\concept{setup}
