% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_path_set}
\alias{restez_path_set}
\title{Set restez path}
\usage{
restez_path_set(filepath)
}
\arguments{
\item{filepath}{character, valid filepath to the folder where the
database should be stored.}
}
\description{
Specify the filepath for the local GenBank database.
}
\details{
Adds 'restez_path' to options(). In this path
the folder 'restez' will be created and all downloaded and
database files will be stored there.
}
\examples{
\dontrun{
library(restez)
restez_path_set(filepath = 'path/to/where/you/want/files/to/download')
}
}
\seealso{
Other setup: \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
\concept{setup}
