% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{set_database_filepath}
\alias{set_database_filepath}
\title{Set database filepath}
\usage{
set_database_filepath(filepath)
}
\arguments{
\item{filepath}{character, valid filepath to the folder where the
database should be stored.}
}
\description{
Specify the filepath for the local GenBank database.
}
\details{
Adds 'restez_database_filepath' to options(). In this path
the database will be created and called 'restez_database'
}
