% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{set_restez_path}
\alias{set_restez_path}
\title{Set restez path}
\usage{
set_restez_path(filepath)
}
\arguments{
\item{filepath}{character, valid filepath to the folder where the
database should be stored.}
}
\description{
Specify the filepath for the local GenBank database.
}
\details{
Adds 'restez_path' to options(). In this path
the folder 'restez' will be created and all downloaded and
database files will be stored there.
}
\seealso{
Other setup: \code{\link{create_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}}
}
