% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{status_check}
\alias{status_check}
\title{Check restez status}
\usage{
status_check()
}
\description{
Report to console current setup status of restez.
Determines if the restez path is set, how many downloaded files
there are, if there is an SQL database.
}
\examples{
library(restez)
fp <- tempdir()
restez_path_set(filepath = fp)
demo_db_create()
status_check()
db_delete()
status_check()
}
\seealso{
Other setup: \code{\link{db_create}},
  \code{\link{db_delete}}, \code{\link{db_download}},
  \code{\link{demo_db_create}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}}
}
