% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{unset_restez_path}
\alias{unset_restez_path}
\title{Unset restez path}
\usage{
unset_restez_path()
}
\description{
Set the restez path to NULL
}
\seealso{
Other setup: \code{\link{check_status}},
  \code{\link{create_database}},
  \code{\link{create_demo_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}}
}
