% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-tools.R
\name{add_to_database}
\alias{add_to_database}
\title{Add to database}
\usage{
add_to_database(df, database)
}
\arguments{
\item{df}{Records data.frame}

\item{database}{Database name}
}
\description{
Add records data.frame to SQLite database.
}
