% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-get-tools.R
\name{count_db_ids}
\alias{count_db_ids}
\title{Return the number of ids}
\usage{
count_db_ids(db = "nucleotide")
}
\arguments{
\item{db}{character, database name}
}
\value{
integer
}
\description{
Return the number of ids in a user's restez database.
}
\details{
Requires an open connection. If no connection or db 0 is returned.
}
\examples{
library(restez)
restez_path_set(filepath = tempdir())
demo_db_create(n = 5)
restez_connect()
(count_db_ids())

# delete demo after example
db_delete(everything = TRUE)
}
\seealso{
Other database: \code{\link{db_create}},
  \code{\link{db_delete}}, \code{\link{db_download}},
  \code{\link{demo_db_create}}, \code{\link{is_in_db}},
  \code{\link{list_db_ids}}
}
\concept{database}
