% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_database}
\alias{create_database}
\title{Create database}
\usage{
create_database(db_type = "nucleotide", overwrite = FALSE)
}
\arguments{
\item{db_type}{character, database type}

\item{overwrite}{T/F, overwrite files already in database?}
}
\description{
Checks for downloaded .seq.tar files,
decompresses and then adds the files to a local SQL
database.
}
\examples{
\dontrun{
library(restez)
set_restez_path(filepath = 'path/for/downloads')
download_genbank()
create_database()
}
}
\seealso{
Other setup: \code{\link{check_status}},
  \code{\link{create_demo_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}},
  \code{\link{unset_restez_path}}
}
