% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_demo_database}
\alias{create_demo_database}
\title{Create demo database}
\usage{
create_demo_database(db_type = "nucleotide", n = 100)
}
\arguments{
\item{db_type}{character, database type}

\item{n}{integer, number of mock sequences}
}
\description{
Creates a local mock SQL database
from package test data for demonstration purposes.
No internet connection required.
}
