% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{create_demo_database}
\alias{create_demo_database}
\title{Create demo database}
\usage{
create_demo_database(db_type = "nucleotide", n = 100)
}
\arguments{
\item{db_type}{character, database type}

\item{n}{integer, number of mock sequences}
}
\description{
Creates a local mock SQL database
from package test data for demonstration purposes.
No internet connection required.
}
\examples{
library(restez)
# set the restez path to a temporary dir
set_restez_path(filepath = tempdir())
# create demo database
create_demo_database(n = 100)
# in the demo, IDs are 'demo_1', 'demo_2' ...
(get_sequence(id = 'demo_1'))
}
\seealso{
Other setup: \code{\link{create_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}}
}
