% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_create}
\alias{db_create}
\title{Create database}
\usage{
db_create(db_type = "nucleotide", overwrite = FALSE)
}
\arguments{
\item{db_type}{character, database type}

\item{overwrite}{T/F, overwrite files already in database?}
}
\description{
Checks for downloaded .seq.tar files,
decompresses and then adds the files to a local SQL
database.
}
\examples{
\dontrun{
library(restez)
restez_path_set(filepath = 'path/for/downloads')
db_download()
db_create()
}
}
\seealso{
Other setup: \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{status_check}}
}
