% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_create}
\alias{db_create}
\title{Create new NCBI database}
\usage{
db_create(db_type = "nucleotide", min_length = 0, max_length = NULL)
}
\arguments{
\item{db_type}{character, database type}

\item{min_length}{Minimum sequence length, default 0.}

\item{max_length}{Maximum sequence length, default NULL.}
}
\description{
Create a new local SQL database from downloaded files.
Currently only GenBank/nucleotide/nuccore database is supported.
}
\details{
All .seq.gz files are added to the database. A user can specify sequence
limit sizes for those sequences to be added to the database -- smaller
databases are faster to  search and is best to limit the database size if
possible.

This function will not overwrite a pre-exisitng database. Old databases must
be deleted before a new one can be created. Use \code{\link{db_delete}} with
everything=FALSE to delete an SQL database.
}
\examples{
\dontrun{
library(restez)
restez_path_set(filepath = 'path/for/downloads')
db_download()
restez_connect()
db_create()
restez_disconnect()
}
}
\seealso{
Other setup: \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
