% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{db_create}
\alias{db_create}
\title{Create new NCBI database}
\usage{
db_create(db_type = "nucleotide", overwrite = FALSE, min_length = 0,
  max_length = NULL)
}
\arguments{
\item{db_type}{character, database type}

\item{overwrite}{T/F, overwrite files already in database?}

\item{min_length}{Minimum sequence length, default 0.}

\item{max_length}{Maximum sequence length, default NULL.}
}
\description{
Create a new local SQL database from downloaded files.
Currently only GenBank/nucleotide/nuccore database is supported.
}
\details{
Decompresses any .seq.tar files to create .seq files, .seq.tar files
are deleted after decompression.

All .seq files are added to the database. A user can specify sequence limit
sizes for those sequences to be added to the database -- smaller databases
are faster to  search and is best to limit the database size if possible.

Whenever this function is run, the old database is deleted and a new one
created.

If overwrite=TRUE, any .seq file with the same name as the newly downloaded
.seq.tar files will be overwritten.
}
\examples{
\dontrun{
library(restez)
restez_path_set(filepath = 'path/for/downloads')
db_download()
db_create()
}
}
\seealso{
Other setup: \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
