% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-setup-tools.R
\name{db_delete}
\alias{db_delete}
\title{Delete database}
\usage{
db_delete(everything = FALSE)
}
\arguments{
\item{everything}{T/F, delete the whole restez folder as well?}
}
\description{
Delete the local SQL database and/or restez folder.
}
\details{
Any connected database will be automatically disconnected.
}
\examples{
library(restez)
fp <- tempdir()
restez_path_set(filepath = fp)
demo_db_create(n = 10)
db_delete(everything = FALSE)
# Will not run: gb_sequence_get(id = 'demo_1')
# only the SQL database is deleted
db_delete(everything = TRUE)
# Now returns NULL
(restez_path_get())
}
\seealso{
Other database: \code{\link{count_db_ids}},
  \code{\link{db_create}}, \code{\link{db_download}},
  \code{\link{demo_db_create}}, \code{\link{is_in_db}},
  \code{\link{list_db_ids}}
}
\concept{database}
