% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{delete_database}
\alias{delete_database}
\title{Delete database}
\usage{
delete_database()
}
\description{
Delete the local SQL database.
}
\seealso{
Other setup: \code{\link{create_database}},
  \code{\link{create_demo_database}},
  \code{\link{download_genbank}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}}
}
