% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{download_genbank}
\alias{download_genbank}
\title{Download GenBank}
\usage{
download_genbank(overwrite = FALSE)
}
\arguments{
\item{overwrite}{T/F, overwrite pre-existing downloaded files?}
}
\description{
Download .seq.tar files from the latest GenBank
release. The user interacitvely selects the parts of
GenBank to download (e.g. primates, plants, bacteria ...)
}
\details{
The downloaded files will appear in the restez filepath under
downloads.
}
\examples{
\dontrun{
library(restez)
set_restez_path(filepath = 'path/for/downloads')
download_genbank()
}
}
\seealso{
Other setup: \code{\link{create_database}},
  \code{\link{create_demo_database}},
  \code{\link{delete_database}},
  \code{\link{get_restez_path}},
  \code{\link{set_restez_path}}
}
