% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-tools.R
\name{extract_by_keyword}
\alias{extract_by_keyword}
\title{Extract by keyword}
\usage{
extract_by_keyword(record, keyword, end_pattern = "\\n")
}
\arguments{
\item{record}{GenBank record in text format, character}

\item{keyword}{Keyword, character}

\item{end_pattern}{REGEX pattern indicating the point to
stop extraction, character}
}
\value{
character
}
\description{
Search through GenBank record for a keyword and
return text up to the end_pattern.
}
\details{
The keyword should be any of the capitalized elements
in a GenBank record (e.g. LOCUS, DESCRIPTION, ACCESSION).
The end_pattern depends on how much of the selected element
a user wants returned. By default, the extraction will stop
at the next newline.
}
