% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{gbrelease_log}
\alias{gbrelease_log}
\title{Log the GenBank release number in the restez path}
\usage{
gbrelease_log(release)
}
\arguments{
\item{release}{GenBank release number, character}
}
\description{
This function is called whenever db_download is run. It logs the
GB release number in the 'gb_release.txt' in the user's restez path.
The log is to help users keep track of whether their database if out of date.
}
\seealso{
Other private: \code{\link{add_rcrd_log}},
  \code{\link{cat_line}}, \code{\link{char}},
  \code{\link{check_connection}}, \code{\link{cleanup}},
  \code{\link{connection_get}},
  \code{\link{custom_download2}},
  \code{\link{custom_download}},
  \code{\link{db_nrows_get}},
  \code{\link{db_sqlngths_get}},
  \code{\link{db_sqlngths_log}}, \code{\link{dir_size}},
  \code{\link{download_cmd_generate}},
  \code{\link{dwnld_path_get}},
  \code{\link{dwnld_rcrd_log}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_patterns}},
  \code{\link{extract_clean_sequence}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_inforecpart}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_seqrecpart}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{filename_log}},
  \code{\link{flatfile_read}}, \code{\link{gb_df_create}},
  \code{\link{gb_df_generate}}, \code{\link{gb_sql_add}},
  \code{\link{gb_sql_query}},
  \code{\link{gbrelease_check}},
  \code{\link{gbrelease_get}}, \code{\link{handle_run}},
  \code{\link{identify_downloadable_files}},
  \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
  \code{\link{last_entry_get}},
  \code{\link{latest_genbank_release_notes}},
  \code{\link{latest_genbank_release}},
  \code{\link{line_spinner}},
  \code{\link{message_missing}}, \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{readme_log}},
  \code{\link{restez_path_check}}, \code{\link{restez_rl}},
  \code{\link{rhandle_generate}},
  \code{\link{seshinfo_log}}, \code{\link{setup}},
  \code{\link{slctn_get}}, \code{\link{slctn_log}},
  \code{\link{sql_path_get}}, \code{\link{stat}},
  \code{\link{testdatadir_get}}
}
