% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_definition}
\alias{get_definition}
\title{Get definition}
\usage{
get_definition(id)
}
\arguments{
\item{id}{sequence accession ID(s)}
}
\value{
list of definitions
}
\description{
Return the definition line
for an accession ID.
}
