% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_definition}
\alias{get_definition}
\title{Get definition}
\usage{
get_definition(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
named vector of definitions
}
\description{
Return the definition line
for an accession ID.
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database()
def <- get_definition(id = 'demo_1')
defs <- get_definition(id = c('demo_1', 'demo_2'))

# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_fasta}},
  \code{\link{get_organism}}, \code{\link{get_record}},
  \code{\link{get_sequence}}, \code{\link{get_version}},
  \code{\link{is_in_db}}, \code{\link{list_db_ids}}
}
