% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_definition}
\alias{get_definition}
\title{Get definition}
\usage{
get_definition(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return definition from GenBank record
}
