% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_organism}
\alias{get_organism}
\title{Get organism}
\usage{
get_organism(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return organism name from GenBank record
}
