% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_organism}
\alias{get_organism}
\title{Get organism}
\usage{
get_organism(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
named vector of definitions
}
\description{
Return the organism name
for an accession ID.
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database()
org <- get_organism(id = 'demo_1')
orgs <- get_organism(id = c('demo_1', 'demo_2'))

# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_fasta}}, \code{\link{get_record}},
  \code{\link{get_sequence}}, \code{\link{is_in_db}},
  \code{\link{list_db_ids}}
}
