% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{get_restez_path}
\alias{get_restez_path}
\title{Get restez path}
\usage{
get_restez_path()
}
\value{
character
}
\description{
Return filepath to where the restez
database is stored.
}
\seealso{
Other setup: \code{\link{create_database}},
  \code{\link{delete_database}},
  \code{\link{download_genbank}},
  \code{\link{set_restez_path}}
}
