% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_sequence}
\alias{get_sequence}
\title{Get sequence}
\usage{
get_sequence(id)
}
\arguments{
\item{id}{character, sequence accession ID(s)}
}
\value{
list of sequences
}
\description{
Return the sequence(s) for a record(s)
from the accession ID(s).
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_organism}}, \code{\link{get_record}},
  \code{\link{list_db_ids}}
}
