% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_sequence}
\alias{get_sequence}
\title{Get sequence}
\usage{
get_sequence(id)
}
\arguments{
\item{id}{sequence accession ID}
}
\value{
character
}
\description{
Return the sequence for a record
from the accession ID.
}
