% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_sequence}
\alias{get_sequence}
\title{Get sequence}
\usage{
get_sequence(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return sequecne from GenBank record
}
