% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{get_version}
\alias{get_version}
\title{Get version}
\usage{
get_version(record)
}
\arguments{
\item{record}{GenBank record in text format, character}
}
\value{
character
}
\description{
Return version ID from GenBank record
}
