% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-tools.R
\name{identify_downloadable_files}
\alias{identify_downloadable_files}
\title{Identify downloadable files}
\usage{
identify_downloadable_files(release_notes)
}
\arguments{
\item{release_notes}{character, GenBank release notes}
}
\value{
data.frame
}
\description{
Searches through the release notes
for a GenBank release to find all listed .seq files.
Returns a data.frame for all .seq files and their
description.
}
\seealso{
Other private: \code{\link{check_connection}},
  \code{\link{cleanup}}, \code{\link{connection_get}},
  \code{\link{custom_download}},
  \code{\link{dwnld_path_get}},
  \code{\link{entrez_fasta_get}},
  \code{\link{entrez_gb_get}},
  \code{\link{extract_accession}},
  \code{\link{extract_by_keyword}},
  \code{\link{extract_definition}},
  \code{\link{extract_features}},
  \code{\link{extract_keywords}},
  \code{\link{extract_locus}},
  \code{\link{extract_organism}},
  \code{\link{extract_sequence}},
  \code{\link{extract_version}},
  \code{\link{file_download}}, \code{\link{flatfile_read}},
  \code{\link{gb_df_create}}, \code{\link{gb_df_generate}},
  \code{\link{gb_sql_add}}, \code{\link{gb_sql_query}},
  \code{\link{identify_latest_genbank_release_notes}},
  \code{\link{mock_def}},
  \code{\link{mock_gb_df_generate}},
  \code{\link{mock_org}}, \code{\link{mock_rec}},
  \code{\link{mock_seq}}, \code{\link{restez_path_check}},
  \code{\link{restez_rl}},
  \code{\link{save_download_details}}, \code{\link{setup}},
  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
}
