% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{is_in_db}
\alias{is_in_db}
\title{Is in db}
\usage{
is_in_db(id, db = "nucleotide")
}
\arguments{
\item{id}{character, sequence accession ID(s)}

\item{db}{character, database name}
}
\value{
named vector of booleans
}
\description{
Determine whether an id(s)
is/are present in a database.
}
\examples{
library(restez)
# set the restez path to a temporary dir
set_restez_path(filepath = tempdir())
# create demo database
create_demo_database(n = 100)
# in the demo, IDs are 'demo_1', 'demo_2' ...
ids <- c('thisisnotanid', 'demo_1', 'demo_2')
(is_in_db(id = ids))

# delete demo after example
delete_database()
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_fasta}}, \code{\link{get_organism}},
  \code{\link{get_record}}, \code{\link{get_sequence}},
  \code{\link{get_version}}, \code{\link{list_db_ids}}
}
