% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{list_db_ids}
\alias{list_db_ids}
\title{List database IDs}
\usage{
list_db_ids(db)
}
\arguments{
\item{db}{character, database name}
}
\value{
vector of characters
}
\description{
Return a vector of all IDs in
a database.
}
