% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-tools.R
\name{list_db_ids}
\alias{list_db_ids}
\title{List database IDs}
\usage{
list_db_ids(db = "nucleotide")
}
\arguments{
\item{db}{character, database name}
}
\value{
vector of characters
}
\description{
Return a vector of all IDs in
a database.
}
\details{
Warning: can return very large vectors
for large databases.
}
\examples{
library(restez)
set_restez_path(filepath = tempdir())
create_demo_database(n = 10)
# Warning: not recommended for real databases
#  with potentially millions of IDs
all_ids <- list_db_ids()
# What shall we do with these IDs?
# ... how about make a mock fasta file
seqs <- get_sequence(id = all_ids)
defs <- get_definition(id = all_ids)
# paste together
fasta_seqs <- paste0('>', defs, '\\n', seqs)
fasta_file <- paste0(fasta_seqs, collapse = '\\n')
cat(fasta_file)
}
\seealso{
Other get: \code{\link{get_definition}},
  \code{\link{get_organism}}, \code{\link{get_record}},
  \code{\link{get_sequence}}
}
