% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-tools.R
\name{restez_connect}
\alias{restez_connect}
\title{Connect to the restez database}
\usage{
restez_connect()
}
\description{
Sets a connection to the local database. If database
connection cannot be made, an error is returned.
}
\examples{
library(restez)
fp <- tempdir()
restez_path_set(filepath = fp)
demo_db_create(n = 5)
restez_connect()
restez_status()
restez_disconnect()
db_delete(everything = TRUE)
# Errors
# restez_status()
}
\seealso{
Other setup: \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
\concept{setup}
