% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-tools.R
\name{restez_disconnect}
\alias{restez_disconnect}
\title{Disconnect from restez database}
\usage{
restez_disconnect()
}
\description{
Safely disconnect from the restez connection
}
\examples{
library(restez)
fp <- tempdir()
restez_path_set(filepath = fp)
demo_db_create(n = 5)
restez_connect()
restez_status()
# always remember to disconnect from a database when you've finished
restez_disconnect()
db_delete(everything = TRUE)
# Errors
# restez_status()
}
\seealso{
Other setup: \code{\link{restez_connect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
\concept{setup}
