% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_path_unset}
\alias{restez_path_unset}
\title{Unset restez path}
\usage{
restez_path_unset()
}
\description{
Set the restez path to NULL
}
\details{
Any connected database will be automatically disconnected.
}
\seealso{
Other setup: \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_ready}}, \code{\link{restez_status}}
}
\concept{setup}
