% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-tools.R
\name{restez_ready}
\alias{restez_ready}
\title{Is restez ready?}
\usage{
restez_ready()
}
\value{
Logical
}
\description{
Returns TRUE if a restez SQL database is available, connected
and has data. Use restez_status() for more information.
}
\examples{
library(restez)
fp <- tempdir()
restez_path_set(filepath = fp)
restez_connect()
demo_db_create()
(restez_ready())
restez_disconnect()
db_delete()
(restez_ready())
}
\seealso{
Other setup: \code{\link{db_create}},
  \code{\link{db_delete}}, \code{\link{db_download}},
  \code{\link{demo_db_create}},
  \code{\link{restez_connect}},
  \code{\link{restez_disconnect}},
  \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_status}}
}
