% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath-tools.R
\name{restez_ready}
\alias{restez_ready}
\title{Is restez ready?}
\usage{
restez_ready(connection = NULL)
}
\arguments{
\item{connection}{A connection for checking. Default NULL.}
}
\value{
T/F
}
\description{
Returns TRUE if a restez SQL database is available. Use
restez_status() for more information. If a connection is already open,
provide the connection as an argument. The function will then test whether
the conection is set up correctly and not empty.
}
\examples{
library(restez)
fp <- tempdir()
restez_path_set(filepath = fp)
demo_db_create()
(restez_ready())
}
\seealso{
Other setup: \code{\link{connect}},
  \code{\link{db_create}}, \code{\link{db_delete}},
  \code{\link{db_download}}, \code{\link{demo_db_create}},
  \code{\link{disconnect}}, \code{\link{restez_path_get}},
  \code{\link{restez_path_set}},
  \code{\link{restez_path_unset}},
  \code{\link{restez_status}}
}
